# public/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: Colorado 14er API
  description: >
    #### Allows users to save and access Colorado 14ers
basePath: /api/v1
parameters:
  range:
    name: range
    in: path
    description: The name of the range
    type: string
    required: true
definitions:
  Links:
    type: object
  PeakCollection:
    type: object
    properties:
      links:
        $ref: '#/definitions/Links'
      data:
        type: array
        items:
          $ref: '#/definitions/PeakResource'
  PeakDocument:
    type: object
    properties:
      links:
        $ref: '#/definitions/Links'
      data:
        $ref: '#/definitions/PeakResource'
  PeakResource:
    type: object
    properties:
      type:
        type: string
        enum:
        - peak
      id:
        name: id
        in: path
        description: The id of the peak
        type: integer
      attributes:
        $ref: '#/definitions/Peak'
      links:
        $ref: '#/definitions/Links'
  Peak:
    type: object
    properties:
      peak_name:
        type: string
      range:
        type: string
      rank:
        type: integer
      elevation:
        type: integer
      latitude:
        type: string
      longitude:
        type: string
  Error:
    type: object
    properties:
      error:
        type: array
        items:
          type: object
          properties:
            status:
              type: string
            source:
              type: object
              properties:
                pointer:
                  type: string
            title:
              type: string
            details:
              type: string
paths:
  '/peaks':
    get:
      description: >
        Lists all peaks
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/PeakCollection'
    post:
      description: >
        Creates a new peak
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/PeakDocument'
        '400':
          description: Validation error
          schema:
            $ref: '#/definitions/Error'
      parameters:
        - name: Payload
          in: body
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    enum:
                    - peak
                  attributes:
                    type: object
                    properties:
                      peak_name:
                        type: string
                      range:
                        type: string
                      rank:
                        type: integer
                      elevation:
                        type: integer
                      latitude:
                        type: string
                      longitude:
                        type: string
  '/peaks/{id}':
    get:
      parameters:
        - $ref: '#/definitions/PeakResource/properties/id'
      description: >
        Lists unique peak
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/PeakCollection'
  '/peaks/ranges/{range}':
    get:
      parameters:
        - $ref: '#/parameters/range'
      description: >
        Lists all peaks for a given range
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/PeakCollection'
  '/peaks/ranges':
    get:
      description: >
        Lists the unique ranges
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              links:
                $ref: '#/definitions/Links'
              data:
                type: array
                items:
                  type: string